cambios reportes asterisk

+++ application/models/mod_gestion.php
@@ -1,7 +1,7 @@
 <?php if ( ! defined('BASEPATH')) exit('No se permite el acceso directo al script');
 
 class mod_gestion extends CI_Model{
-	
+
 	function reporte_gestion($obj){
 
 		$this->db->where('id >', 1);
@@ -69,15 +69,15 @@ class mod_gestion extends CI_Model{
 
 		foreach($row_tipo as $rt){
 			//detalle de las acciones
-			
+
 			$this->db->from('core_cliente_gestion ccg');
 			$this->db->where("ccg.fecha between '".$objb['inicio']." 00:00:01' AND '".$objb['termino']." 23:59:59'");
 			$this->db->where('ccg.idvendedor', $obja['idvendedor']);
 			$this->db->where('ccg.idgestion', $rt['id']);
 			$this->db->group_by('ccg.rbd, ccg.idgestion');
 			$query = $this->db->get();
-			$row = $query->result_array();	
-			
+			$row = $query->result_array();
+
 			array_push($data, array("tipo"=>$rt, "detalle"=>$row));
 		}
 		return $data;
@@ -90,7 +90,7 @@ class mod_gestion extends CI_Model{
 		$this->db->join('core_clientes_asignaciones cca', 'cu.ID=cca.idusuario', 'left');
 		$this->db->group_by('cu.ID');
 		$this->db->order_by('visible desc, NOM_EJECUTIVO');
-		
+
 		if($obj['id']>0){
 			$this->db->where('cu.ID', $obj['id']);
 		}
@@ -101,7 +101,7 @@ class mod_gestion extends CI_Model{
 	}
 
 	function update_usuarios($obj){
-		$datos = array("NOM_EJECUTIVO"=>$obj['nombre'], "EMAIL"=>$obj['email'], 
+		$datos = array("NOM_EJECUTIVO"=>$obj['nombre'], "EMAIL"=>$obj['email'],
 						"LOGIN"=>$obj['loggin'],"PASSWORD"=>$obj['password'],
 						"visible"=>$obj['visible'], "ANEXO"=>$obj['anexo']);
 		$this->db->where('ID', $obj['id']);
@@ -114,23 +114,49 @@ class mod_gestion extends CI_Model{
 		return $this->db->insert('asignacion_realizada', $datos);
 	}
 
-	function crear_asignacion($obj){
+	function filtrar_vendedores_carga(){
+		$this->db->select('cu.ID, cu.LOGIN');
+		$this->db->from('asignacion a');
+		$this->db->join('core_usuarios cu', 'a.vendedor = cu.LOGIN', 'left');
+		$this->db->group_by('cu.id');
+		$query = $this->db->get();
+		return $query->result_array();
+	}
+
+	function crear_asignacion($obj, $carga){
+
 		$this->db->truncate('asignacion');
 
-		foreach($obj['rbd'] as $rbd){
-
-			$datos = array("rbd"=>$rbd, "vendedor"=>$obj['usuario']);
+		foreach($obj as $col){
+			$datos = array("rbd"=>$col['rbd'], "vendedor"=>$col['usuario']);
 			$this->db->insert('asignacion', $datos);
-
 		}
 
-		$this->db->where('idusuario', $obj['usuario']);
-		$this->db->delete('core_clientes_asignaciones');
+		$usuario = $this->filtrar_vendedores_carga();
+		$contador = 1;
+		foreach($usuario as $u){
 
-		$sql="insert into core_clientes_asignaciones (idusuario, idcliente, idrbd, idasignador)
-			  select vendedor, rbd, rbd, '168' from asignacion";
+			if($carga == 1){
 
-		return $this->db->query($sql);
+				$sql="insert into core_clientes_asignaciones (idusuario, idcliente, idrbd, idasignador)
+					  select '".$u['ID']."', rbd, rbd, '168' from asignacion where vendedor = '".$u['LOGIN']."'";
+
+				$this->db->query($sql);
+
+			}elseif($carga == 0){
+
+				$this->db->where('idusuario', $u['ID']);
+				$this->db->delete('core_clientes_asignaciones');
+
+				$sql2="insert into core_clientes_asignaciones (idusuario, idcliente, idrbd, idasignador)
+				  select '".$u['ID']."', rbd, rbd, '168' from asignacion  where vendedor = '".$u['LOGIN']."'";
+			 	$this->db->query($sql2);
+			}
+
+			$contador++;
+		}
+
+		return $contador;
 
 	}
